Done with primary payload systems

+++ payload/aprs.h
@@ -10,6 +10,6 @@ namespace aprs
     inline audio_buffer_pool_t* abp = nullptr;
     inline std::string transmit_data;
 
-    void transmitAllData(const PayloadState& state, int apogee);
+    void transmitAllData(const PayloadState& state, double apogee);
     void transmitData(audio_buffer_pool_t* audio_buffer_pool, const std::string& data);
 }
